feat-chore: moved setup_logger to logging package
