+++ autobots/core/logging/setup_logger.py
